git_sync/requirements.txt
gitpython>=2.13
slacker>=0.9.42